Reduce cpu utilisation of sphere-leds
Previously, a background goroutine woke up every second to traverse a data structure and wrote the results into the file system.
It did this to support the "flash" function, even if no LEDs were flashing.

Now, the background thread wakes up every second and checks a boolean to workout if flashing is enabled. If yes, then
it does what it needs to do. Otherwise, it just goes back to sleep.

Signed-off-by: Jon Seymour <jon.seymour@ninjablocks.com>